function_modules/toonify/decode-function/requirements.txt
boto3
redis
ffmpeg-python
requests
